requirements.txt
python-telegram-bot
weather-api
tinydb
lxml
cssselect
requests